Add files via upload
This paper, developed during my MSc in Quantitative Finance at POLIMI Graduate School of Management in Milan, offers a critical and interdisciplinary analysis of the energy impact of Bitcoin mining. It explores the evolution of mining technologies, the exponential growth in electricity consumption and proposes possible pathways toward a more sustainable and responsible crypto ecosystem.